requirements.txt
click==8.1.8
python-dotenv==1.0.1
# Editable install with no version control (slack-migrate==0.1)
-e /Users/florianplank/Projects/slack-migration
slack_bolt==1.22.0
slack_sdk==3.34.0
tabulate==0.9.0
